Hecho el ej6 de herencias

+++ java/src/organizacion/empresa/paginaweb/componentes/CPU.java
@@ -1,7 +1,7 @@
 package organizacion.empresa.paginaweb.componentes;
 
 public class CPU extends Componente{
-    public CPU(int stock, float precioventa, String modelo, String nomfab) {
+    public CPU(int stock, Double precioventa, String modelo, String nomfab) {
         super(stock, precioventa, modelo, nomfab);
     }
 }
