+++ java/src/organizacion/empresa/paginaweb/componentes/dispositivo/entrada/Mouse.java
@@ -1,7 +1,7 @@
 package organizacion.empresa.paginaweb.componentes.dispositivo.entrada;
 
 public class Mouse extends DispositivosEntrada{
-    public Mouse(int stock, float precioventa, String modelo, String nomfab, int cantpuertos, String tipoconector) {
+    public Mouse(int stock, double precioventa, String modelo, String nomfab, int cantpuertos, String tipoconector) {
         super(stock, precioventa, modelo, nomfab, cantpuertos, tipoconector);
     }
 }
