+++ java/src/organizacion/empresa/paginaweb/componentes/dispositivo/salida/DispositivosSalida.java
@@ -5,7 +5,7 @@ import organizacion.empresa.paginaweb.componentes.Componente;
 public class DispositivosSalida extends Componente {
     private int cantpuertos;
 
-    public DispositivosSalida(int stock, float precioventa, String modelo, String nomfab, int cantpuertos) {
+    public DispositivosSalida(int stock, double precioventa, String modelo, String nomfab, int cantpuertos) {
         super(stock, precioventa, modelo, nomfab);
         this.cantpuertos = cantpuertos;
     }
