+++ java/src/organizacion/empresa/paginaweb/componentes/dispositivo/salida/Impresora.java
@@ -3,7 +3,7 @@ package organizacion.empresa.paginaweb.componentes.dispositivo.salida;
 public class Impresora extends DispositivosSalida {
 private String metodoimpr;
 
-    public Impresora(int stock, float precioventa, String modelo, String nomfab, int cantpuertos, String metodoimpr) {
+    public Impresora(int stock, double precioventa, String modelo, String nomfab, int cantpuertos, String metodoimpr) {
         super(stock, precioventa, modelo, nomfab, cantpuertos);
         this.metodoimpr = metodoimpr;
     }
